Refactoring: - moved expression related functions to new unit - renamed integer types
git-svn-id: https://asmotor.googlecode.com/svn/trunk@57 fe0e4675-b9e4-17b0-9a9e-715403563874

+++ util/amitime.c
@@ -26,7 +26,7 @@ void time_Init(void)
 	}
 }
 
-ULONG time_GetMicroSeconds(void)
+uint32_t time_GetMicroSeconds(void)
 {
 	struct timeval tv;
